requirements.txt
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
google-auth==2.12.0
google-auth-oauthlib==0.5.3
gspread==5.5.0
httplib2==0.20.4
idna==3.4
numpy==1.23.3
oauth2client==4.1.3
oauthlib==3.2.1
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.5.0
pyparsing==3.0.9
pypng==0.20220715.0
PyQRCode==1.2.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
urllib3==1.26.12
